requirements.txt
requests
djangorestframework
Django==3.0.4
psycopg2-binary==2.8.4
Pillow==7.1.1
gunicorn==20.0.4
django_heroku
redis
Celery